requirements.txt
boto3
awscli
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit
langchain_groq
langchain_chroma
ipykernel
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
duckduckgo-search
validators==0.28.1
pinecone
langchain_pinecone
langchain_community
